+++ TheatreProject.OrderAPI/Models/OrderHeader.cs
@@ -9,8 +9,8 @@ public class OrderHeader
     public Guid Id { get; set; }
     public string? UserId { get; set; }
     public string? CouponCode { get; set; }
-    public double GrandTotal { get; set; }
-    public double DiscountTotal { get; set; }
+    public decimal GrandTotal { get; set; }
+    public decimal DiscountTotal { get; set; }
     public string? FirstName { get; set; }
     public string? LastName { get; set; }
     public DateTime OrderTime { get; set; }
